+++ R/netcomb.R
@@ -25,6 +25,8 @@
 #' @param nchar.comps A numeric defining the minimum number of
 #'   characters used to create unique names for components (see
 #'   Details).
+#' @param func.inverse R function used to calculate the pseudoinverse
+#'   of the Laplacian matrix L (see \code{\link{netmeta}}).
 #' @param warn.deprecated A logical indicating whether warnings should
 #'   be printed if deprecated arguments are used.
 #' @param \dots Additional arguments (to catch deprecated arguments).
@@ -358,6 +360,8 @@ netcomb <- function(x,
                     details.chkident = FALSE,
                     nchar.comps = x$nchar.trts,
                     ##
+                    func.inverse = invmat,
+                    ##
                     warn.deprecated = gs("warn.deprecated"),
                     ...) {
   
@@ -371,7 +375,7 @@ netcomb <- function(x,
   ##
   x <- updateversion(x)
   ##
-  chkchar(sep.comps, nchar = 1, length = 1)
+  chkchar(sep.comps, nchar = 0:1, length = 1)
   ##
   if (!is.null(tau.preset))
     chknumeric(tau.preset, min = 0, length = 1)
@@ -423,7 +427,11 @@ netcomb <- function(x,
     ##
     ## Create C-matrix from netmeta object
     ##
-    C.matrix <- createC(x, sep.comps, inactive)
+    if (sep.comps == "")
+      C.matrix <- createC(x, "...this_is_not_a_separator...", inactive)
+    else
+      C.matrix <- createC(x, sep.comps, inactive)
+    ##
     inactive <- attr(C.matrix, "inactive")
     C.matrix <- as.matrix(C.matrix)[trts, , drop = FALSE]
   }
@@ -477,7 +485,8 @@ netcomb <- function(x,
   c <- ncol(C.matrix) # number of components
   
   
-  p0 <- prepare(TE, seTE, treat1, treat2, studlab)
+  p0 <- prepare(TE, seTE, treat1, treat2, studlab,
+                func.inverse = func.inverse)
   ##
   o <- order(p0$order)
   ##
@@ -573,7 +582,8 @@ netcomb <- function(x,
   ##
   ## Random effects models
   ##
-  p1 <- prepare(TE, seTE, treat1, treat2, studlab, tau)
+  p1 <- prepare(TE, seTE, treat1, treat2, studlab, tau, invmat)
+  ##
   res.r <- nma.additive(p1$TE[o], p1$weights[o], p1$studlab[o],
                         p1$treat1[o], p1$treat2[o], x$level.ma,
                         X.matrix, C.matrix, B.matrix,
@@ -747,6 +757,8 @@ netcomb <- function(x,
               nchar.comps = nchar.comps,
               ##
               inactive = inactive,
+              ##
+              func.inverse = deparse(substitute(func.inverse)),
               ##
               backtransf = x$backtransf,
               ##