+++ R/netmetabin.R
@@ -88,6 +88,8 @@
 #'   between treatment labels.
 #' @param nchar.trts A numeric defining the minimum number of
 #'   characters used to create unique treatment names (see Details).
+#' @param func.inverse R function used to calculate the pseudoinverse
+#'   of the Laplacian matrix L (see \code{\link{netmeta}}).
 #' @param backtransf A logical indicating whether results should be
 #'   back transformed in printouts and forest plots. If
 #'   \code{backtransf = TRUE}, results for \code{sm = "OR"} are
@@ -405,6 +407,8 @@ netmetabin <- function(event1, n1, event2, n2,
                        ##
                        sep.trts = ":",
                        nchar.trts = 666,
+                       ##
+                       func.inverse = invmat,
                        ##
                        backtransf = gs("backtransf"),
                        ##
@@ -1970,6 +1974,8 @@ netmetabin <- function(event1, n1, event2, n2,
               sep.trts = sep.trts,
               nchar.trts = nchar.trts,
               ##
+              func.inverse = deparse(substitute(func.inverse)),
+              ##
               backtransf = backtransf,
               ##
               title = title,
@@ -2065,7 +2071,8 @@ netmetabin <- function(event1, n1, event2, n2,
                 rep(1, nrow(dat.wide)),
                 dat.wide$treat1,
                 dat.wide$treat2,
-                dat.wide$studlab)
+                dat.wide$studlab,
+                func.inverse = func.inverse)
   ##
   tdata <- data.frame(studies = p0$studlab, narms = p0$narms,
                       stringsAsFactors = FALSE)