+++ R/print.netmeta.R
@@ -274,7 +274,7 @@ print.netmeta <- function(x,
   if (header)
     matitle(x)
   ##  
-  if (reference.group != "" & missing(all.treatments))
+  if (reference.group != "" & is.null(all.treatments))
     all.treatments <- FALSE
   ##
   if (reference.group != "")
@@ -284,7 +284,8 @@ print.netmeta <- function(x,
     cat(paste("Number of studies: k = ", k, "\n", sep = ""))
     cat(paste0("Number of pairwise comparisons: m = ", m, "\n"))
     if (!is.null(x$n.trts))
-      cat(paste0("Number of observations: o = ", sum(x$n.trts, na.rm = TRUE),
+      cat(paste0("Number of observations: o = ",
+                 round(sum(x$n.trts, na.rm = TRUE), 1),
                  "\n"))
     cat(paste0("Number of treatments: n = ", n, "\n"))
     if (!oldversion)