+++ R/print.summary.netmeta.R
@@ -283,7 +283,7 @@ print.summary.netmeta <- function(x,
     if (details) {
       multiarm <- any(x$x$narms > 2)
       cat(paste("Original data",
-                ifelse(multiarm,
+                ifelse(multiarm & (fixed | random),
                        paste(" (with adjusted standard errors for",
                              "multi-arm studies)"),
                        ""),