+++ R/rankogram.R
@@ -7,11 +7,10 @@
 #'
 #' @param x An object of class \code{\link{netmeta}}.
 #' @param nsim Number of simulations.
-#' @param fixed A logical indicating to compute ranking
-#'   probabilities and SUCRAs for the fixed effects (common effects)
-#'   model.
-#' @param random A logical indicating to compute ranking
-#'   probabilities and SUCRAs for the random effects model.
+#' @param fixed A logical indicating to compute ranking probabilities
+#'   and SUCRAs for the fixed effects (common effects) model.
+#' @param random A logical indicating to compute ranking probabilities
+#'   and SUCRAs for the random effects model.
 #' @param small.values A character string specifying whether small
 #'   treatment effects indicate a beneficial (\code{"good"}) or
 #'   harmful (\code{"bad"}) effect, can be abbreviated.
@@ -83,6 +82,7 @@
 #' @rdname rankogram
 #' @export rankogram
 
+
 rankogram <- function(x, nsim = 1000,
                       fixed = x$fixed, random = x$random,
                       small.values = x$small.values,
@@ -137,155 +137,23 @@ rankogram <- function(x, nsim = 1000,
   ## (3) Resampling to calculate ranking probabilites and SUCRAs
   ##
   ##
-  resampling <-
-    lapply(1:nsim,
-           function(y) matrix(NA,
-                              nrow = nrow(x$TE.random),
-                              ncol = ncol(x$TE.random),
-                              dimnames = list(rownames(x$TE.random),
-                                              colnames(x$TE.random))))
+  ranking.fixed  <- ranking.matrix.fixed  <- cumrank.matrix.fixed  <- NULL
+  ranking.random <- ranking.matrix.random <- rank.cum.random <- NULL
   ##
-  rearrange <- function(y, simul) {
-    resampling[[y]][lower.tri(resampling[[y]])] <-  simul[y, ]
-    resampling[[y]] <- t(resampling[[y]])
-    resampling[[y]][lower.tri(resampling[[y]])] <- -simul[y, ]
-    resampling
-  }
-  ##  
   if (fixed) {
-    simul.fixed <- mvtnorm::rmvnorm(nsim,
-                                    t(x$TE.fixed)[lower.tri(x$TE.fixed)],
-                                    x$Cov.fixed)
+    res.f <- ranksampling(x, nsim, "fixed", small.values)
     ##
-    resampling.fixed <- lapply(1:nsim,
-                               function(y)
-                                 rearrange(y, simul = simul.fixed)[[y]])
-    ##
-    if(small.values == "good") {
-      rankings.fixed <- lapply(1:nsim,
-                                function(y)
-                                  rank(rowSums(resampling.fixed[[y]]> 0,
-                                               na.rm = TRUE)))
-      sortedtreats.fixed <- order(x$TE.fixed[, 1])
-    }
-    else if (small.values == "bad") {
-      rankings.fixed <- lapply(1:nsim,
-                               function(y)
-                               (x$n + 1) -
-                               rank(rowSums(resampling.fixed[[y]] > 0,
-                                            na.rm = TRUE)))
-      sortedtreats.fixed <- order(x$TE.fixed[1, ])
-    }
-    ##
-    df.fixed <-
-      data.frame(x = mapply(function(i){
-        (rankings.fixed[[i]][sortedtreats.fixed])}, 1:nsim))
-    ##
-    ## Ranking matrix
-    ##
-    ranking.matrix.fixed <-
-      mapply(function(x){
-        mapply(function(y) {
-          sum((df.fixed[x, ] == y) / nsim)
-        },
-        1:length(rownames(df.fixed)))
-      }, 1:length(rownames(df.fixed)))
-    ##
-    rownames(ranking.matrix.fixed) <- rownames(df.fixed)
-    colnames(ranking.matrix.fixed) <- 1:x$n
-    ##
-    ## Cumulative ranking matrix
-    ##
-    rank.cum.fixed <- t(mapply(function(i)
-      cumsum(ranking.matrix.fixed[i,]), 1:nrow(ranking.matrix.fixed)))
-    ##
-    dimnames(rank.cum.fixed) <- list(rownames(ranking.matrix.fixed), 1:x$n)
-    ##
-    ## SUCRAs
-    ##
-    ranking.fixed <- mapply(function(i)
-      sum(rank.cum.fixed[i, -length(rank.cum.fixed[i, ])]) /
-      (nrow(rank.cum.fixed) - 1), 
-      1:nrow(ranking.matrix.fixed))
-    ##
-    names(ranking.fixed) <- rownames(ranking.matrix.fixed)
-    ##
-    ranking.fixed <- ranking.fixed[x$trts]
+    ranking.fixed <- res.f$ranking
+    ranking.matrix.fixed <- res.f$rankogram
+    cumrank.matrix.fixed <- res.f$cumrank
   }
-  ##  
+  ##
   if (random) {
-    simul.random <- mvtnorm::rmvnorm(nsim,
-                                     t(x$TE.random)[lower.tri(x$TE.random)],
-                                     x$Cov.random)
+    res.r <- ranksampling(x, nsim, "random", small.values)
     ##
-    resampling.random <- lapply(1:nsim,
-                                function(y)
-                                  rearrange(y, simul = simul.random)[[y]])
-    ##
-    if(small.values == "good") {
-      rankings.random <- lapply(1:nsim,
-                                function(y)
-                                  rank(rowSums(resampling.random[[y]]> 0,
-                                               na.rm = TRUE)))
-      sortedtreats.random <- order(x$TE.random[, 1])
-    }
-    else if (small.values == "bad") {
-      rankings.random <- lapply(1:nsim,
-                                function(y)
-                                (x$n + 1) -
-                                rank(rowSums(resampling.random[[y]] > 0,
-                                             na.rm = TRUE)))
-      sortedtreats.random <- order(x$TE.random[1, ])
-    }
-    ##
-    df.random <-
-      data.frame(x = mapply(function(i){
-        (rankings.random[[i]][sortedtreats.random])}, 1:nsim))
-    ##
-    ## Ranking matrix
-    ##
-    ranking.matrix.random <-
-      mapply(function(x){
-        mapply(function(y) {
-          sum((df.random[x, ] == y) / nsim)
-        },
-        1:length(rownames(df.random)))
-      }, 1:length(rownames(df.random)))
-    ##
-    rownames(ranking.matrix.random) <- rownames(df.random)
-    colnames(ranking.matrix.random) <- 1:x$n
-    ##
-    ## Cumulative ranking matrix
-    ##
-    rank.cum.random <- t(mapply(function(i)
-      cumsum(ranking.matrix.random[i,]), 1:nrow(ranking.matrix.random)))
-    ##
-    dimnames(rank.cum.random) <- list(rownames(ranking.matrix.random), 1:x$n)
-    ##
-    ## SUCRAs
-    ##
-    ranking.random <- mapply(function(i)
-      sum(rank.cum.random[i, -length(rank.cum.random[i, ])]) /
-      (nrow(rank.cum.random) - 1), 
-      1:nrow(ranking.matrix.random))
-    ##
-    names(ranking.random) <- rownames(ranking.matrix.random)
-    ##
-    ranking.random <- ranking.random[x$trts]
-    ##
-    cumrank.matrix.random <- t(apply(ranking.matrix.random, 1, cumsum))
-  }
-  ##
-  if (!fixed) {
-    ranking.matrix.fixed <- NULL
-    ranking.fixed <- NULL
-    rank.cum.fixed <- NULL
-  }
-  ##
-  if (!random) {
-    ranking.matrix.random <- NULL
-    ranking.random <- NULL
-    rank.cum.random <- NULL
+    ranking.random <- res.r$ranking
+    ranking.matrix.random <- res.r$rankogram
+    rank.cum.random <- res.r$cumrank
   }
   
   
@@ -294,19 +162,21 @@ rankogram <- function(x, nsim = 1000,
   ## (4) Create rankogram object
   ##
   ##
-  res <- list(ranking.matrix.fixed = ranking.matrix.fixed,
-              ranking.fixed = ranking.fixed,
-              ranking.matrix.random = ranking.matrix.random,
-              ranking.random = ranking.random,
+  res <- list(ranking.fixed = ranking.fixed,
+              ranking.matrix.fixed = ranking.matrix.fixed,
+              cumrank.matrix.fixed = cumrank.matrix.fixed,
               ##
-              cumrank.matrix.fixed = rank.cum.fixed,
+              ranking.random = ranking.random,
+              ranking.matrix.random = ranking.matrix.random,
               cumrank.matrix.random = rank.cum.random,
               ##
               nsim = nsim,
+              ##
               fixed = fixed,
               random = random,
               small.values = small.values,
               cumulative.rankprob = cumulative.rankprob,
+              ##
               nchar.trts = nchar.trts,
               x = x,
               version = packageDescription("netmeta")$Version